skip existing files, continue if error

+++ download.js
@@ -10,27 +10,39 @@ const assets = JSON.parse(data);
 // download i.gltf and i.image
 // save them both to ./assets/i.id (opensea unique id)
 
+// TODO:
+// - if the asset already exists, skip it
+// - if remove slashes from name
+
 async function download() {
   console.log(assets.length);
   for (const asset of assets) {
     if (!asset.gltf || !asset.image) continue;
-    const gltf = await fetch(asset.gltf);
-    const image = await fetch(asset.image);
-    const gltfBuffer = await gltf.buffer();
-    const imageBuffer = await image.buffer();
-    const name = asset.name.replace(/\s/g, "_");
+    try {
+      const gltf = await fetch(asset.gltf);
+      const image = await fetch(asset.image);
+      const gltfBuffer = await gltf.buffer();
+      const imageBuffer = await image.buffer();
+      // remove slashes from name an replace spaces with underscores
+      const name = asset.name.replace(/\//g, "").replace(/ /g, "_");
 
-    fs.mkdirSync(path.join(__dirname, "assets", `${asset.id}`));
-    fs.writeFileSync(
-      path.join(__dirname, "assets", `${asset.id}`, `${name}.gltf`),
-      gltfBuffer
-    );
-    fs.writeFileSync(
-      path.join(__dirname, "assets", `${asset.id}`, `${name}.png`),
-      imageBuffer
-    );
-
-    console.log(`Downloaded ${name}`);
+      // create directory if it doesn't exist
+      if (!fs.existsSync(path.join(__dirname, "assets", `${asset.id}`))) {
+        fs.mkdirSync(path.join(__dirname, "assets", `${asset.id}`));
+        fs.writeFileSync(
+          path.join(__dirname, "assets", `${asset.id}`, `${name}.gltf`),
+          gltfBuffer
+        );
+        fs.writeFileSync(
+          path.join(__dirname, "assets", `${asset.id}`, `${name}.png`),
+          imageBuffer
+        );
+      }
+      console.log(`Downloaded ${name}`);
+    } catch (e) {
+      console.log(e);
+      continue;
+    }
   }
 }
